+++ rs/consensus/utils/Cargo.toml
@@ -18,6 +18,7 @@ ic-replicated-state = { path = "../../replicated_state" }
 ic-types = { path = "../../types/types" }
 prometheus = { workspace = true }
 rand = { workspace = true }
+rayon = { workspace = true }
 slog = { workspace = true }
 thiserror = { workspace = true }
