+++ rs/state_machine_tests/src/lib.rs
@@ -11,11 +11,9 @@ use ic_config::{
     state_manager::LsmtConfig,
     subnet_config::SubnetConfig,
 };
-use ic_consensus::consensus::{
-    MAX_CONSENSUS_THREADS, build_thread_pool, payload_builder::PayloadBuilderImpl,
-};
+use ic_consensus::consensus::payload_builder::PayloadBuilderImpl;
 use ic_consensus_cup_utils::make_registry_cup;
-use ic_consensus_utils::crypto::SignVerify;
+use ic_consensus_utils::{MAX_CONSENSUS_THREADS, build_thread_pool, crypto::SignVerify};
 use ic_crypto_test_utils_crypto_returning_ok::CryptoReturningOk;
 use ic_crypto_test_utils_ni_dkg::{
     SecretKeyBytes, dummy_initial_dkg_transcript_with_master_key, sign_message,
